refactor(result): split class in three separate files
