+++ TagsCloudResult/TagsCloudVisualization.WFApp/Common/PictureBoxImageHolder.cs
@@ -1,5 +1,6 @@
 ﻿using System.Drawing.Imaging;
 using TagsCloudVisualization.Common;
+using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.WFApp.Common;
