+++ TagsCloudResult/TagsCloudVisualization/CloudLayouters/CircularCloudLayouter.cs
@@ -1,5 +1,6 @@
 ﻿using System.Drawing;
 using TagsCloudVisualization.Common;
+using TagsCloudVisualization.Common.ResultOf;
 using TagsCloudVisualization.Extensions;
 using TagsCloudVisualization.PointsProviders;
