+++ TagsCloudResult/TagsCloudVisualization/Common/ResultOf/None.cs
@@ -1,0 +1,8 @@
+﻿namespace TagsCloudVisualization.Common.ResultOf;
+
+public class None
+{
+    private None()
+    {
+    }
+}
