+++ TagsCloudResult/TagsCloudVisualization/Common/ResultOf/Result.cs
@@ -1,0 +1,24 @@
+﻿namespace TagsCloudVisualization.Common.ResultOf;
+
+public struct Result<T>
+{
+    public Result(string error, T value = default(T))
+    {
+        Error = error;
+        Value = value;
+    }
+        
+    public static implicit operator Result<T>(T v)
+    {
+        return Result.Ok(v);
+    }
+
+    public string Error { get; }
+    internal T Value { get; }
+    public T GetValueOrThrow()
+    {
+        if (IsSuccess) return Value;
+        throw new InvalidOperationException($"No value. Only Error {Error}");
+    }
+    public bool IsSuccess => Error == null;
+}
