+++ TagsCloudResult/TagsCloudVisualization/Extensions/PointExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Drawing;
 using System.Numerics;
+using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.Extensions;
