+++ TagsCloudResult/TagsCloudVisualization/PointsProviders/ArchimedeanSpiralPointsProvider.cs
@@ -1,5 +1,6 @@
 ﻿using System.Drawing;
 using TagsCloudVisualization.Common;
+using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.PointsProviders;
