+++ TagsCloudResult/TagsCloudVisualization/TextReaders/DocTextReader.cs
@@ -1,6 +1,7 @@
 ﻿using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Wordprocessing;
 using TagsCloudVisualization.Common;
+using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.TextReaders;
