+++ TagsCloudResult/TagsCloudVisualization/WordsProcessors/SimpleWordsProcessor.cs
@@ -1,4 +1,5 @@
 ﻿using TagsCloudVisualization.Common;
+using TagsCloudVisualization.Common.ResultOf;
 
 namespace TagsCloudVisualization.WordsProcessors;
